requirements.txt
telebot~=0.0.5
tensorflow~=2.16.1
requests~=2.31.0
python-dotenv~=1.0.1
opencv-python~=4.9.0.80
numpy~=1.26.4
pandas~=2.2.1
keras~=3.1.1
torch~=2.2.1
torchvision~=0.17.1
tqdm~=4.64.1